init mostly building
had to do a few things so far: 1. rough in packages from the monorepo this was extracted from using local paths; 2. delete a bunch of node-in-app code which seems to only be used for sentry; 3. tweak some metro and babel build configs; 4. switch to yarn 1.x which is more standard in RN world; 5. fix the missing bridging header on iOS; 6. bunch more i suspect